lockstat: export data in python expression
Hi,

I'm writing the tool for parsing and arranging the data of /proc/lock_stat
in python. But I soon found that lock_stat is hard to parse, so I wrote
this patch to let lock_stat output its information in python expression.

This patch implements the ioctl() for lockstat. via this ioctl(), readers can
change the format of printing. If it is not called lockstat prints information
in traditional style. If it is called and specified cmd as 0 and arg as 1,
readers can get the information in python expression.

The big benefit of printing in pytohn expression is that it can be parsed by
eval() of python.

Sample of output:
{ 'version': 0.3,}	# header

{ 'name': 'pm_qos_lock', 'writer': [        3572153 ,       3573984,            0.11,          372.87,      3333033.22,        23292427 ,      31255420,            0.10,          291.25,     18881384.46, ], 'reader': None, 'contention': [[3573984, 0xffffffff8106e4ea, 'pm_qos_request+0x1a/0x70'], ], 'contending': [[3573984, 0xffffffff8106e4ea, 'pm_qos_request+0x1a/0x70'], ] }

{ 'name': 'xtime_lock', 'writer': [        1798546 ,       1798795,            0.27,           10.92,      4252980.74,         7116071 ,      14729524,            0.00,         1000.00,     30979660.87, ], 'reader': None, 'contention': [[1798795, 0xffffffff81077dec, 'tick_do_update_jiffies64+0x3c/0xd0'], ], 'contending': [[1798795, 0xffffffff81077dec, 'tick_do_update_jiffies64+0x3c/0xd0'], ] }

...

And,

f = open('/proc/lock_stat')
fcntl.ioctl(f.fileno(), 0, 1)	# specify format

header = eval(f.readline())	# the first line is header

for line in f.readlines():
    lock = eval(line)		# eval each line

this is the everything of parsing, really easy. I think exporting data in
eval()able style is efficient way in procfs and sysfs. How do you think?

BTW, this patch is too python centric. Exporting in more neutral style like
S-expression might be good.

Signed-off-by: Hitoshi Mitake <mitake@dcl.info.waseda.ac.jp>
Cc: Peter Zijlstra <a.p.zijlstra@chello.nl>
Cc: Paul Mackerras <paulus@samba.org>
Cc: Ingo Molnar <mingo@elte.hu>
Cc: Arnaldo Carvalho de Melo <acme@ghostprotocols.net>
Cc: Steven Rostedt <rostedt@goodmis.org>