tools: add new headers for futex and atomic operations
This patch adds two new header files in tools/ directory. The purpose
of adding these two headers is supporting futextest and perf bench
futex in kernel tree. But these can be utilized by other purposes
because they are generalized enough.

This patch produces checkpatch warns because of typedefing and volatile
declaration. But I think these warns are not essential problems